src/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
SQLAlchemy==2.0.48
pydantic-settings==2.13.1
slowapi==0.1.9
bcrypt==5.0.0
email-validator==2.3.0
python-jose[cryptography]==3.5.0
httpx==0.28.1

# mysql library
pymysql==1.1.1